ID-1072 Lazily Load Authorization URLs from Bond
