+++ src/profile/external-identities/FrameworkServiceLink.tsx
@@ -1,0 +1,37 @@
+import { ClickableProps } from '@terra-ui-packages/components';
+import React, { ReactNode } from 'react';
+import { ButtonPrimary, Link } from 'src/components/common';
+import { icon } from 'src/components/icons';
+import { Ajax } from 'src/libs/ajax';
+import { withErrorReporting } from 'src/libs/error';
+import * as Utils from 'src/libs/utils';
+
+interface FrameworkServiceLinkProps extends ClickableProps {
+  linkText: string;
+  providerKey: string;
+  redirectUrl: string;
+  button?: boolean;
+}
+
+export const FrameworkServiceLink = (props: FrameworkServiceLinkProps): ReactNode => {
+  const { linkText, providerKey, redirectUrl, button = false, ...clickableProps } = props;
+  const Component = button ? ButtonPrimary : Link;
+  const style = button ? {} : { display: 'inline-flex', alignItems: 'center' };
+  return (
+    <Component
+      onClick={async () => {
+        const loadAuthUrlAndRedirect = withErrorReporting('Error getting Fence Link', async () => {
+          const result = await Ajax().User.getFenceAuthUrl(providerKey, redirectUrl);
+          window.open(result.url, Utils.newTabLinkProps.target, Utils.newTabLinkProps.rel);
+        });
+        loadAuthUrlAndRedirect();
+      }}
+      style={style}
+      /* eslint-disable-next-line react/jsx-props-no-spreading */
+      {...clickableProps}
+    >
+      {linkText}
+      {icon('pop-out', { size: 12, style: { marginLeft: '0.2rem' } })}
+    </Component>
+  );
+};
